feat: Add script to resize and add logo to images
This commit adds a new script, resizeAndAddLogo.py, which can be used to resize images and add a logo to them. The script takes all .png and .jpg files in the working directory, resizes them if necessary, and pastes a logo image in the bottom-right corner. The modified images are saved in the output_images folder. This script is useful for adding branding or watermarks to images.